+++ include/GafferUI/Tool.h
@@ -40,6 +40,7 @@
 #include "GafferUI/Export.h"
 #include "GafferUI/TypeIds.h"
 
+#include "Gaffer/Container.h"
 #include "Gaffer/Node.h"
 #include "Gaffer/TypedPlug.h"
 
@@ -76,6 +77,7 @@ class GAFFERUI_API Tool : public Gaffer::Node
 
 	public :
 
+		Tool( View *view, const std::string &name = defaultName<Tool>() );
 		~Tool() override;
 
 		GAFFER_NODE_DECLARE_TYPE( GafferUI::Tool, ToolTypeId, Gaffer::Node );
@@ -87,6 +89,11 @@ class GAFFERUI_API Tool : public Gaffer::Node
 		/// is currently active.
 		Gaffer::BoolPlug *activePlug();
 		const Gaffer::BoolPlug *activePlug() const;
+
+		/// The Tool constructor automatically parents the tool to
+		/// the`View::toolsContainer()`. After that, the tool may not be
+		/// reparented.
+		bool acceptsParent( const GraphComponent *potentialParent ) const override;
 
 		/// @name Factory
 		///////////////////////////////////////////////////////////////////
@@ -102,8 +109,6 @@ class GAFFERUI_API Tool : public Gaffer::Node
 		//@}
 
 	protected :
-
-		Tool( View *view, const std::string &name = defaultName<Tool>() );
 
 		template<typename ToolType, typename ViewType>
 		struct ToolDescription
@@ -126,11 +131,15 @@ class GAFFERUI_API Tool : public Gaffer::Node
 
 	private :
 
-		View *m_view;
+		/// \todo Remove.
+		View *m_unused;
 
 		static size_t g_firstPlugIndex;
 
 };
+
+using ToolContainer = Gaffer::Container<Gaffer::Node, Tool>;
+IE_CORE_DECLAREPTR( ToolContainer );
 
 } // namespace GafferUI